+++ src/meico/chroma/Key.java
@@ -6,16 +6,15 @@ package meico.chroma;
  */
 
 public class Key {
-
-    public double[] referenceFrequencies;          // the reference frequencies
-    public boolean octaveModulo;                   // is the chroma freature represent pitch classes (true) or absolute pitches (false)
+    private double[] referenceFrequencies;          // the reference frequencies
+    private boolean octaveModulo;                   // is the chroma freature represent pitch classes (true) or absolute pitches (false)
 
     /**
      * default constructor
      * creates key for equal tempered tuning over concert a 440 Hz
      */
     public Key() {
-        this(new double[]{261.6, 277.2, 293.7, 311.1, 329.6, 349.2, 370.0, 392.0, 415.3, 440.0, 466.2, 493.9}, true);
+        this(new double[]{261.63, 277.18, 293.66, 311.13, 329.63, 349.23, 369.99, 392.0, 415.3, 440.0, 466.16, 493.88}, true);
     }
 
     /**
@@ -26,7 +25,6 @@ public class Key {
     public Key(double[] referenceFrequencies, boolean octaveModulo) {
         this.octaveModulo = octaveModulo;
         this.referenceFrequencies = referenceFrequencies;
-        this.normalize();
     }
 
     /**
@@ -38,24 +36,18 @@ public class Key {
     }
 
     /**
+     * this getter returns the size/length of the vector of reference frequencies
+     * @return
+     */
+    public int getSize() {
+        return this.referenceFrequencies.length;
+    }
+
+    /**
      * a getter to return whether the chroma features represent pitch classes (true) or absolute pitches (false)
      * @return
      */
     public boolean getOctaveModulo() {
         return this.octaveModulo;
-    }
-
-    /**
-     * This methhod scales the reference frequencies down so that the lowest frequency is 1.
-     * However, this method has only an effect if octaveModulo is true!
-     */
-    public void normalize() {
-        if (!this.octaveModulo) return;                                                         // if this key represents absolute pitches, normalization should not be applied
-
-        double[] normalizedFreqs = new double[this.referenceFrequencies.length];
-        for (int i = this.referenceFrequencies.length - 1; i >= 0; --i) {                       // for each reference frequency
-            normalizedFreqs[i] = this.referenceFrequencies[i] / this.referenceFrequencies[0];   // devide it by the first/lowest frequency
-        }
-        this.referenceFrequencies = normalizedFreqs;
     }
 }